SIH/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
paho-mqtt==2.1.0
redis==5.0.8
